refactor(controller): add abstractclass; move publish trigger into common.release
