+++ src/dashboard/apigateway/apigateway/biz/stage.py
@@ -22,7 +22,7 @@ from typing import Any, Dict, Optional
 from django.db import transaction
 
 from apigateway.biz.release import ReleaseHandler
-from apigateway.controller.tasks.syncing import trigger_gateway_publish
+from apigateway.common.release.publish import trigger_gateway_publish
 from apigateway.core.constants import DEFAULT_BACKEND_NAME, DEFAULT_STAGE_NAME, PublishSourceEnum, StageStatusEnum
 from apigateway.core.models import Backend, BackendConfig, MicroGateway, Release, Stage
 from apigateway.utils.time import now_datetime
