+++ src/dashboard/apigateway/apigateway/controller/management/commands/sync_releases_to_shared_micro_gateway.py
@@ -20,7 +20,7 @@ from typing import List, Optional
 
 from django.core.management.base import BaseCommand, CommandError
 
-from apigateway.controller.tasks import syncing
+from apigateway.common.release import publish
 from apigateway.core.constants import PublishSourceEnum
 from apigateway.core.models import Gateway
 
@@ -44,7 +44,7 @@ class Command(BaseCommand):
         for gateway in gateways:
             print(f"syncing release for gateway {gateway.name} ...")
             # publish_id=-1 标识cli同步网关发布操作，方便operator过滤不上报
-            ok = syncing.trigger_gateway_publish(
+            ok = publish.trigger_gateway_publish(
                 PublishSourceEnum.CLI_SYNC, author="cli", gateway_id=gateway.id, is_sync=True
             )
             if not ok:
